+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/IndexTask.java
@@ -108,7 +108,6 @@ import org.apache.druid.server.security.AuthorizerMapper;
 import org.apache.druid.timeline.DataSegment;
 import org.apache.druid.timeline.partition.HashBasedNumberedShardSpec;
 import org.apache.druid.timeline.partition.NumberedShardSpec;
-import org.apache.druid.timeline.partition.ShardSpec;
 import org.apache.druid.utils.CircularBuffer;
 import org.codehaus.plexus.util.FileUtils;
 import org.joda.time.Interval;
@@ -1011,39 +1010,6 @@ public class IndexTask extends AbstractBatchIndexTask implements ChatHandler
       return publishFuture.get();
     } else {
       return publishFuture.get(publishTimeout, TimeUnit.MILLISECONDS);
-    }
-  }
-
-  /**
-   * This class represents a map of (Interval, ShardSpec) and is used for easy shardSpec generation.
-   */
-  static class ShardSpecs
-  {
-    private final Map<Interval, List<ShardSpec>> map;
-    private Granularity queryGranularity;
-
-    ShardSpecs(final Map<Interval, List<ShardSpec>> map, Granularity queryGranularity)
-    {
-      this.map = map;
-      this.queryGranularity = queryGranularity;
-    }
-
-    /**
-     * Return a shardSpec for the given interval and input row.
-     *
-     * @param interval interval for shardSpec
-     * @param row      input row
-     *
-     * @return a shardSpec
-     */
-    ShardSpec getShardSpec(Interval interval, InputRow row)
-    {
-      final List<ShardSpec> shardSpecs = map.get(interval);
-      if (shardSpecs == null || shardSpecs.isEmpty()) {
-        throw new ISE("Failed to get shardSpec for interval[%s]", interval);
-      }
-      final long truncatedTimestamp = queryGranularity.bucketStart(row.getTimestamp()).getMillis();
-      return shardSpecs.get(0).getLookup(shardSpecs).getShardSpec(truncatedTimestamp, row);
     }
   }
